split test suite
