+++ .gitignore
@@ -52,6 +52,8 @@ UpgradeLog*.htm
 *.vspx
 *.sap
 
+# ignore mpi.c generated by make
+mpi.c
 
 # ignore stuff generated by "make manual" and "make poster"
 *.aux
